requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==7.0.6
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==4.4.2
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
executing==2.2.0
faiss-cpu==1.13.2
fastjsonschema==2.21.1
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.47
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
imageio==2.25.1
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-openai==1.2.0
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
lark==1.2.2
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mistune==3.1.3
moviepy==1.0.3
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.20.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.1
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.1
pillow==11.3.0
platformdirs==4.3.8
proglog==0.1.12
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==7.34.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==24.0.0
pycparser==2.22
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydeck==0.9.2
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.4
pypdf==6.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
qiskit==2.1.1
qiskit-aer==0.17.1
referencing==0.36.2
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
rustworkx==0.16.0
scipy==1.16.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.49
stack-data==0.6.3
stevedore==5.4.1
streamlit==1.56.0
tenacity==9.1.4
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2026.1
uri-template==1.3.0
urllib3==2.5.0
uuid_utils==0.14.1
uv==0.9.24
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.23.0
yt-dlp==2026.3.17
zstandard==0.25.0
